+++ backend/flow/src/Flow/Migrations.hs
@@ -118,6 +118,46 @@ createTableFlowInstanceKeyValueStore = Migration
       }
   }
 
+createTableFlowUserAuthConfigs :: MonadDB m => Migration m
+createTableFlowUserAuthConfigs = Migration
+  { mgrTableName = "flow_user_auth_configs"
+  , mgrFrom      = 0
+  , mgrAction    =
+    StandardMigration . createTable True $ tblTable
+      { tblName        = "flow_user_auth_configs"
+      , tblVersion     = 1
+      , tblColumns     =
+        [ tblColumn { colName = "instance_id", colType = UuidT, colNullable = False }
+        , tblColumn { colName = "key", colType = TextT, colNullable = False }
+        , tblColumn { colName     = "auth_to_view_provider"
+                    , colType     = TextT
+                    , colNullable = True
+                    }
+        , tblColumn { colName     = "auth_to_view_max_failures"
+                    , colType     = IntegerT
+                    , colNullable = True
+                    }
+        , tblColumn { colName     = "auth_to_view_archived_provider"
+                    , colType     = TextT
+                    , colNullable = True
+                    }
+        , tblColumn { colName     = "auth_to_view_archived_max_failures"
+                    , colType     = IntegerT
+                    , colNullable = True
+                    }
+        ]
+      , tblPrimaryKey  = pkOnColumns ["instance_id", "key"]
+      , tblIndexes     = []
+      , tblForeignKeys = [ (fkOnColumns ["instance_id", "key"]
+                                        "flow_instance_key_value_store"
+                                        ["instance_id", "key"]
+                           )
+                             { fkOnDelete = ForeignKeyCascade
+                             }
+                         ]
+      }
+  }
+
 createTableFlowInstanceSignatories :: MonadDB m => Migration m
 createTableFlowInstanceSignatories = Migration
   { mgrTableName = "flow_instance_signatories"