+++ backend/test/src/FlowTests.hs
@@ -6,6 +6,7 @@ import TestKontra
 import qualified Flow.AggregatorTest as Aggregator
 import qualified Flow.AuthenticationTest as Authentication
 import qualified Flow.DocumentCheckerTest as DocumentChecker
+import qualified Flow.EIDTest as EID
 import qualified Flow.HighTongueTest as HighTongue
 import qualified Flow.IntegrationTest as Integration
 import qualified Flow.VariableCollectorTest as VariableCollector
@@ -16,6 +17,7 @@ flowTests env = testGroup
   [ Aggregator.tests
   , Authentication.tests env
   , DocumentChecker.tests
+  , EID.tests env
   , HighTongue.tests
   , Integration.tests env
   , VariableCollector.tests